+++ bin/nanocld/src/services/cargo.rs
@@ -396,7 +396,7 @@ mod tests {
             name: test_cargo.to_owned(),
             container: bollard_next::container::Config {
               image: Some(
-                "ghcr.io/nxthat/nanocl-get-started:latest".to_owned(),
+                "ghcr.io/next-hat/nanocl-get-started:latest".to_owned(),
               ),
               ..Default::default()
             },
@@ -415,7 +415,7 @@ mod tests {
       assert_eq!(cargo.namespace_name, "global", "Invalid cargo namespace");
       assert_eq!(
         cargo.spec.container.image,
-        Some("ghcr.io/nxthat/nanocl-get-started:latest".to_owned())
+        Some("ghcr.io/next-hat/nanocl-get-started:latest".to_owned())
       );
     }
     let mut res = client
@@ -486,7 +486,9 @@ mod tests {
         Some(&CargoSpecPartial {
           name: main_test_cargo.to_owned(),
           container: bollard_next::container::Config {
-            image: Some("ghcr.io/nxthat/nanocl-get-started:latest".to_owned()),
+            image: Some(
+              "ghcr.io/next-hat/nanocl-get-started:latest".to_owned(),
+            ),
             env: Some(vec!["TEST=1".to_owned()]),
             ..Default::default()
           },
@@ -501,7 +503,7 @@ mod tests {
     assert_eq!(patch_response.namespace_name, "global");
     assert_eq!(
       patch_response.spec.container.image,
-      Some("ghcr.io/nxthat/nanocl-get-started:latest".to_owned())
+      Some("ghcr.io/next-hat/nanocl-get-started:latest".to_owned())
     );
     assert_eq!(
       patch_response.spec.container.env,
@@ -569,7 +571,9 @@ mod tests {
         Some(&CargoSpecPartial {
           name: CARGO_NAME.to_owned(),
           container: bollard_next::container::Config {
-            image: Some("ghcr.io/nxthat/nanocl-get-started:latest".to_owned()),
+            image: Some(
+              "ghcr.io/next-hat/nanocl-get-started:latest".to_owned(),
+            ),
             ..Default::default()
           },
           ..Default::default()