+++ scripts/build_images.sh
@@ -12,8 +12,8 @@ for project in ./bin/*; do
   #   continue
   # fi
   echo "Building ${name}:${version}"
-  docker build -t "ghcr.io/nxthat/$name:$version-nightly" -f $project/Dockerfile .
-  # docker buildx build --builder=$(BUILDER) --platform=linux/amd64,linux/arm64 --tag="ghcr.io/nxthat/$name:$version" -f $project/Dockerfile .
+  docker build -t "ghcr.io/next-hat/$name:$version-nightly" -f $project/Dockerfile .
+  # docker buildx build --builder=$(BUILDER) --platform=linux/amd64,linux/arm64 --tag="ghcr.io/next-hat/$name:$version" -f $project/Dockerfile .
 done
 
 echo "Done!"