+++ scripts/install_dev_image.sh
@@ -7,8 +7,8 @@ mv ubuntu-22.04-minimal-cloudimg-amd64.img tests/ubuntu-22.04-minimal-cloudimg-a
 
 echo "Downloading container images.."
 docker pull cockroachdb/cockroach:v23.1.12
-docker pull ghcr.io/nxthat/metrsd:0.3.2
-docker pull ghcr.io/nxthat/nanocl-get-started:latest
-docker pull ghcr.io/nxthat/nanocl-dev:dev
+docker pull ghcr.io/next-hat/metrsd:0.3.2
+docker pull ghcr.io/next-hat/nanocl-get-started:latest
+docker pull ghcr.io/next-hat/nanocl-dev:dev
 docker build --network host -t ndns:dev -f ./bin/ndns/Dockerfile .
 docker build --network host -t nproxy:dev -f ./bin/nproxy/Dockerfile .